Move game_items::GameItem struct definition to GameItems/Item.h
